+++ src/components/shared/Modal/SideDrawer/SideDrawer.type.ts
@@ -1,19 +1,8 @@
-import { ModalPosition } from './EditFolderMoreModal/EditFolderMoreModal.type';
-import { Folder } from '@/types/forest';
-
 export type StyledProps = {
 	show: boolean;
 };
 
 export type Props = {
-	checkedTreeId?: number;
-	trees: Folder[] | undefined;
 	onModal: boolean;
 	setOnModal: (state: boolean) => void;
-	modalPosition: ModalPosition;
-	onEditMoreModal: boolean;
-	handleEditMoreModalOpen: (event: React.MouseEvent<HTMLElement>) => void;
-	handleEditMoreModalClose: () => void;
-	handleFolderDeleteAlertModalToggle: (state: 'open' | 'close') => void;
-	onClickTreeFolderMoreMenuButton: (treeId: number) => void;
 };
