+++ src/pages/MessageBox/index.tsx
@@ -3,18 +3,13 @@ import { useParams } from 'react-router-dom';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import * as S from './MessageBox.styled';
 import { getMessages, deleteMessage } from '@/apis/messages';
-import { getForest, deleteTree } from '@/apis/forest';
 import { MessagesType } from '@/types/message';
-import { AlertModal, MovingFolderModal, SideDrawer, DeleteAlertModal } from '@/components/shared';
+import { MovingFolderModal, SideDrawer, DeleteAlertModal } from '@/components/shared';
 import { MessageMenu, MessageContent, MakingFruitMenu, BottomButtons } from '@/components/features/MessageBox';
-import { Folder } from '@/types/forest';
-
-import { useRecoilValue } from 'recoil';
-import { myInfoState } from '@/stores/user';
 
 const MessageBox = () => {
 	const { treeId } = useParams();
-	const myInfo = useRecoilValue(myInfoState);
+
 	const queryClient = useQueryClient();
 
 	const { data: messages } = useQuery<MessagesType>(['getMessages', treeId], () => getMessages(treeId));
@@ -25,57 +20,15 @@ const MessageBox = () => {
 		},
 	});
 
-	const { data: trees } = useQuery<Folder[] | undefined>(['getForest', myInfo?.id], () => getForest(myInfo?.id), {
-		enabled: !!myInfo,
-	});
-
-	const treeDeleteMutation = useMutation(deleteTree, {
-		onSuccess: () => {
-			queryClient.invalidateQueries('getForest');
-			handleFolderDeleteAlertModalToggle('close');
-		},
-	});
-
 	const [checkMessages, setCheckMessages] = useState<number[]>([]);
 	const [isEdit, setIsEdit] = useState(false);
 	const [checkMode, setCheckMode] = useState(false);
 	const [isMakingFruit, setIsMakingFruit] = useState(false);
 	const [isMoving, setIsMoving] = useState(false);
 	const [showCheckedMessages, setShowCheckedMessages] = useState(false);
-	const [isOpenDeleteModal, setIsOpenDeleteModal] = useState(false);
+	const [isOpenedMessageDeleteAlertModal, setIsOpenedMessageDeleteAlertModal] = useState(false);
 
 	const [openedDrawer, setOpenedDrawer] = useState(false);
-	const [modalPosition, setModalPosition] = useState({ top: 0, left: 0 });
-	const [onEditMoreModal, setOnEditMoreModal] = useState(false);
-	const [isOpenedMessageDeleteAlertModal, setIsOpenedMessageDeleteAlertModal] = useState(false);
-	const [isOpenedFolderDeleteAlertModal, setIsOpenedFolderDeleteAlertModal] = useState(false);
-	const [checkedTreeId, setCheckedTreeId] = useState<number>();
-
-	const handleEditMoreModalOpen = (event: React.MouseEvent<HTMLElement>) => {
-		const closest = event.currentTarget.closest('a') as HTMLAnchorElement;
-		const rect = closest.getBoundingClientRect();
-		const newPosition = { top: rect.top, left: rect.left + rect.width };
-
-		setModalPosition(newPosition);
-		setOnEditMoreModal(true);
-	};
-
-	const handleEditMoreModalClose = () => {
-		setOnEditMoreModal(false);
-	};
-
-	const handleClickTreeFolderMoreMenuButton = (treeId: number) => {
-		setCheckedTreeId(treeId);
-	};
-
-	const handleFolderDelete = () => {
-		treeDeleteMutation.mutate(checkedTreeId);
-	};
-
-	const handleFolderDeleteAlertModalToggle = (state: 'open' | 'close') => {
-		setIsOpenedFolderDeleteAlertModal(state === 'open');
-		handleEditMoreModalClose();
-	};
 
 	const onToggleCheckMessage = (id: number) => {
 		checkMessages.includes(id)
@@ -128,10 +81,6 @@ const MessageBox = () => {
 			setIsEdit(false);
 			setIsMakingFruit(false);
 		}
-	};
-
-	const handleMessageDeleteAlertModalToggle = (state: 'open' | 'close') => {
-		setIsOpenedMessageDeleteAlertModal(state === 'open');
 	};
 
 	useEffect(() => {
@@ -213,19 +162,6 @@ const MessageBox = () => {
 				)}
 			</S.MessageListContainer>
 
-			<SideDrawer
-				checkedTreeId={checkedTreeId}
-				trees={trees}
-				onModal={openedDrawer}
-				setOnModal={onToggleOpenDrawer}
-				onEditMoreModal={onEditMoreModal}
-				modalPosition={modalPosition}
-				handleEditMoreModalOpen={handleEditMoreModalOpen}
-				handleEditMoreModalClose={handleEditMoreModalClose}
-				handleFolderDeleteAlertModalToggle={handleFolderDeleteAlertModalToggle}
-				onClickTreeFolderMoreMenuButton={handleClickTreeFolderMoreMenuButton}
-			/>
-
 			{isMoving && (
 				<MovingFolderModal
 					isMoving={isMoving}
@@ -239,20 +175,12 @@ const MessageBox = () => {
 					deleteTargetType="message"
 					deleteTarget="메시지"
 					onAlertModal={isOpenedMessageDeleteAlertModal}
-					handleAlertModalToggle={() => handleMessageDeleteAlertModalToggle('close')}
+					handleAlertModalToggle={() => setIsOpenedMessageDeleteAlertModal(false)}
 					handleTargetDelete={deleteMessageHandler}
 				/>
 			)}
 
-			{isOpenedFolderDeleteAlertModal && (
-				<DeleteAlertModal
-					deleteTargetType="folder"
-					deleteTarget="프로젝트A"
-					onAlertModal={isOpenedFolderDeleteAlertModal}
-					handleAlertModalToggle={() => handleFolderDeleteAlertModalToggle('close')}
-					handleTargetDelete={handleFolderDelete}
-				/>
-			)}
+			<SideDrawer onModal={openedDrawer} setOnModal={onToggleOpenDrawer} />
 		</S.Wrapper>
 	);
 };
